My Test Scene with my new Scripts
Scripts include:
+ Room (Spawning random item prefabs in specified locations, destroy
items in the room, alerting the player to change rooms when the room is
entered)
+ Event Deck (draw random or specified events, keep track of events that
have and have not been activate)
+ IEvent (a parent class all events derive from)
+ Player (a temporary class for basic player stats and the room they're
currently in.  Also, the player now has a flashlight)

Events include:
+ BreakerBox (turn on and off lights)
+ CaveInEvent (for now, only activates a gameObject)
+ FireEvent (destroys all items in the specified rooms, spawns a fire
prefab in that room (*BUT* I should change this to a specified location)

My Test Scene demonstrates the breaker box and how to set up the basic
scripts.  However, there's a lot of missing functionality.  I'll need a
day/night cycle to activate events, an official map to properly set up
the scene, and, eventually, official prefabs to replace placeholders.